more work for iros
Former-commit-id: ec8fbaed625f72c7e3d791b2323483c9ed5d667b
Former-commit-id: 35289f7e1090195b02b9b9390142fac33d1f3680 [formerly ab9ec7c6639e77fe121c9051940b05bfbef9145f] [formerly e5b86c32886c517587601450083d7a1f99497991 [formerly 634ac0f9f214fafce7145735f9b3d9f0e90d2d63]]
Former-commit-id: f0e8b4ba76a61cb9d24565ef957b0f8f6f2335a5 [formerly db4d42187db1928dd3ec25b20f15e16936fa12f2]
Former-commit-id: 6f0425833422004184d7f9645227a79524ebf1cd

+++ seagul/envs/classic_control/acrobot.py
@@ -72,11 +72,11 @@ class SGAcroEnv(core.Env):
         self.render_length2 = 1.0
         self.viewer = None
 
-        low = np.array([th1_range[0], th2_range[0], -max_th1dot, -max_th2dot])
-        high = np.array([th1_range[1], th2_range[1], max_th1dot, max_th2dot])
+        low = np.array([th1_range[0], th2_range[0], -max_th1dot, -max_th2dot], dtype=np.float32)
+        high = np.array([th1_range[1], th2_range[1], max_th1dot, max_th2dot], dtype=np.float32)
 
         self.observation_space = spaces.Box(low=low-1, high=high+1, dtype=np.float32)
-        self.action_space = spaces.Box(low=np.array([-max_torque]), high=np.array([max_torque]), dtype=np.float32)
+        self.action_space = spaces.Box(low=np.array([-max_torque],dtype=np.float32), high=np.array([max_torque], dtype=np.float32), dtype=np.float32)
 
         self.t = 0
 
@@ -192,7 +192,7 @@ class SGAcroEnv(core.Env):
 
         d2th = np.linalg.solve(M,(TAU - H - PHI))
         return np.array([th1d, th2d, d2th[0].item(), d2th[1].item()])
-    #
+
     # def _dynamics(self, t, s0, act):
     #     m1 = self.m1
     #     m2 = self.m2
